# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M-PESA Business Loans API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ---------- Storage ----------
...
  /loan-applications:
    get:
      tags: [Loans]
      operationId: listMyLoanApplications
      summary: List the current customer's loan applications
      responses:
        '200':
          description: List of loan applications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanApplication'
    post:
      tags: [Loans]
      operationId: createLoanApplication
      summary: Apply for a business loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationInput'
      responses:
        '201':
          description: Loan application created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanApplication'
  /loan-applications/{id}:
    get:
      tags: [Loans]
      operationId: getMyLoanApplication
      summary: Get one of the current customer's loan applications
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Loan application found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanApplication'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /loans/{id}/repayments:
    get:
      tags: [Loans]
      operationId: getLoanRepaymentSchedule
      summary: Get the repayment schedule for a loan
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Repayment schedule.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Repayment'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ---------- Notifications ----------
...
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: Mark a notification as read
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Notification updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ---------- Admin ----------
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role, accountStatus]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          $ref: '#/components/schemas/UserRole'
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ----- Storage -----
...
    UserRole:
      type: string
      enum: [super_admin, loan_officer, customer]
    AccountStatus:
      type: string
      enum: [active, suspended]

    # ----- Profile -----
...
    LoanApplicationStatus:
      type: string
      enum: [pending, approved, rejected, hold]
    LoanApplication:
      type: object
      required:
        - id
        - customerId
        - amount
        - purpose
        - loanType
        - termMonths
        - status
        - reviewedBy
        - reviewedAt
        - reviewNotes
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        customerId: { type: string }
        amount: { type: string }
        purpose: { type: string }
        loanType: { type: string }
        termMonths: { type: integer }
        status:
          $ref: '#/components/schemas/LoanApplicationStatus'
        reviewedBy: { type: [string, 'null'] }
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewNotes: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LoanApplicationInput:
      type: object
      required: [amount, purpose, loanType, termMonths]
      properties:
        amount:
          type: string
          description: Decimal amount as a string, e.g. "50000.00"
        purpose:
          type: string
          minLength: 1
        loanType:
          type: string
          minLength: 1
        termMonths:
          type: integer
          minimum: 1
          maximum: 60
...
    RepaymentStatus:
      type: string
      enum: [pending, paid, overdue]
    Repayment:
      type: object
      required:
        - id
        - loanId
        - installmentNumber
        - amountDue
        - dueDate
        - amountPaid
        - paidAt
        - status
        - createdAt
      properties:
        id: { type: string }
        loanId: { type: string }
        installmentNumber: { type: integer }
        amountDue: { type: string }
        dueDate: { type: string }
        amountPaid: { type: [string, 'null'] }
        paidAt: { type: [string, 'null'], format: date-time }
        status:
          $ref: '#/components/schemas/RepaymentStatus'
        createdAt: { type: string, format: date-time }

    # ----- Notifications -----
    NotificationChannel:
      type: string
      enum: [sms, email, in_app]
    NotificationStatus:
      type: string
      enum: [sent, pending, failed]
    Notification:
      type: object
      required: [id, userId, channel, title, message, status, read, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        channel:
          $ref: '#/components/schemas/NotificationChannel'
        title: { type: string }
        message: { type: string }
        status:
          $ref: '#/components/schemas/NotificationStatus'
        read: { type: boolean }
        createdAt: { type: string, format: date-time }

    # ----- Admin -----
...