# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  super_admin: 'super_admin',
  loan_officer: 'loan_officer',
  customer: 'customer',
} as const;